server.js modularization, might work idk
